+++ app/routes/components/Toast/changelog.codegen.ts
@@ -6,4 +6,5 @@ export const changelog = ''
   + '- 2025-09-23: Add support for pausing the auto-dismiss timer when hovering\n'
   + '  over a toast & add basic styles.\n'
   + '- 2025-10-05: Code formatting improvement.\n'
+  + '- 2025-10-28: Use `Modal.Raw` to render the toast container.\n'
   ;