+++ src/components/Toast/CHANGELOG.md
@@ -2,3 +2,4 @@
 - 2025-09-23: Add support for pausing the auto-dismiss timer when hovering
   over a toast & add basic styles.
 - 2025-10-05: Code formatting improvement.
+- 2025-10-28: Use `Modal.Raw` to render the toast container.